+++ solid/form/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @einfach/solid-form
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.8
+  - @einfach/solid@0.2.8
+  - @einfach/utils@0.2.6
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.7
+  - @einfach/solid@0.2.7
+  - @einfach/utils@0.2.5
+
 ## 0.2.6
 
 ### Patch Changes