+++ solid/solid/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @einfach/solid
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.8
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/core@0.2.7
+
 ## 0.2.6
 
 ### Patch Changes